+++ src/main/java/com/example/wini/domain/auth/service/AuthService.java
@@ -37,6 +37,6 @@ public class AuthService {
             return memberRepository.save(newMember);
         });
 
-        return tokenService.generateTokens(member);
+        return tokenService.upsertTokens(member);
     }
 }
