+++ src/main/java/com/example/wini/domain/auth/service/TokenService.java
@@ -1,19 +1,15 @@
 package com.example.wini.domain.auth.service;
 
-import static com.example.wini.global.common.constant.SecurityConstants.REFRESH_TOKEN;
 import static com.example.wini.global.common.constant.SecurityConstants.TOKEN_BLACKLIST_CACHE_NAME;
-import static com.example.wini.global.error.exception.ErrorCode.INVALID_REFRESH_TOKEN;
-import static com.example.wini.global.error.exception.ErrorCode.MEMBER_NOT_FOUND;
 
 import com.example.wini.domain.auth.domain.RefreshToken;
 import com.example.wini.domain.auth.dto.response.TokenResponse;
 import com.example.wini.domain.auth.repository.RefreshTokenRepository;
 import com.example.wini.domain.common.util.MemberUtil;
 import com.example.wini.domain.member.domain.Member;
-import com.example.wini.domain.member.repository.MemberRepository;
-import com.example.wini.global.error.exception.CustomException;
 import com.example.wini.global.security.AuthMember;
 import com.example.wini.global.security.JwtProvider;
+import java.util.Optional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.cache.annotation.CachePut;
 import org.springframework.cache.annotation.Cacheable;
@@ -24,50 +20,23 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 public class TokenService {
 
-    private final MemberRepository memberRepository;
     private final RefreshTokenRepository refreshTokenRepository;
     private final JwtProvider jwtProvider;
     private final MemberUtil memberUtil;
 
     @Transactional
-    public TokenResponse generateTokens(Member member) {
-        String accessToken = jwtProvider.generateAccessToken(member);
-        String refreshToken = jwtProvider.generateRefreshToken(member);
-
-        RefreshToken newRefreshTokenEntity = RefreshToken.create(member, jwtProvider.substringToken(refreshToken));
-        refreshTokenRepository.save(newRefreshTokenEntity);
-
-        return TokenResponse.of(accessToken, refreshToken);
-    }
-
-    @Transactional
-    public TokenResponse reissueToken(String refreshTokenHeader) {
-        String refreshToken = jwtProvider.substringToken(refreshTokenHeader);
-        validateRefreshToken(refreshToken);
-
-        return updateAndGenerateTokens(refreshToken);
-    }
-
-    private void validateRefreshToken(String refreshToken) {
-        if (!jwtProvider.validToken(refreshToken, REFRESH_TOKEN)) {
-            throw new CustomException(INVALID_REFRESH_TOKEN);
-        }
-    }
-
-    private TokenResponse updateAndGenerateTokens(String refreshToken) {
-        Long userId = jwtProvider.getMemberId(refreshToken, REFRESH_TOKEN);
-
-        Member member = memberRepository.findById(userId).orElseThrow(() -> new CustomException(MEMBER_NOT_FOUND));
-
-        RefreshToken storedRefreshToken = refreshTokenRepository
-                .findByToken(refreshToken)
-                .orElseThrow(() -> new CustomException(INVALID_REFRESH_TOKEN));
-
+    public TokenResponse upsertTokens(Member member) {
         String newAccessToken = jwtProvider.generateAccessToken(member);
         String newRefreshToken = jwtProvider.generateRefreshToken(member);
 
-        storedRefreshToken.updateToken(newRefreshToken);
-        refreshTokenRepository.save(storedRefreshToken);
+        Optional<RefreshToken> optionalRefreshToken = refreshTokenRepository.findByMemberId(member.getId());
+        if (optionalRefreshToken.isPresent()) {
+            RefreshToken storedToken = optionalRefreshToken.get();
+            storedToken.updateToken(newRefreshToken);
+        } else {
+            RefreshToken newRefreshTokenEntity = RefreshToken.create(member, newRefreshToken);
+            refreshTokenRepository.save(newRefreshTokenEntity);
+        }
 
         return TokenResponse.of(newAccessToken, newRefreshToken);
     }
